refactor(ts): extract generic async Queue from multiagent
